Final Project COMP 2132
In final-project folder, I created an index.html file, a .vscode folder, an images folder, a scripts folder, and a styles folder. I included SCSS with compiled CSS files, a settings.json for the SCSS files, added images for dice game, added some jquery animations, and javascript for the functionality of the code.